+++ docs/changelog.md
@@ -51,6 +51,7 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 - **[BREAKING]** The `warpMouse` property is now a global option and cannot be set per menu anymore. It also does not only affect the Centered Mode anymore, but disables any mouse warping, even close to the screen's edge. Mouse warping is enabled by default.
 - The lazy initialization introduced in 1.8.0 has been made an option which is disabled by default. This means that the menu will be initialized when Kando is started. This should make the first menu open fast again.
 - Slightly reduced the icon size of the default, rainbow labels, and clean circle themes to reduce the amount of icon clipping.
+- When the settings dialog should be shown but is minimized, Kando will now restore the settings dialog so that it is visible again.
 
 #### :bug: Fixed